Fix HACD index passing #89

+++ BulletSharp/demos/ConvexDecompositionDemo/ConvexDecompositionDemo.cs
@@ -230,7 +230,7 @@ namespace ConvexDecompositionDemo
                 {
                     continue;
                 }
-                var triangles = new long[trianglesLen];
+                var triangles = new int[trianglesLen];
 
                 int nVertices = hacd.GetNPointsCH(c);
                 var points = new double[nVertices * 3];
